Randomly picked dataset from the Illinois area from 1979.

Data from the Illinois area from one point from the 32km grid from NOAA from 1979.

Format

A data frame with 365 observations of the following 9 variables.

year

Year (numeric; always 1979)

month

Month (not available; can be computed from JD)

day

Day of the month (not available; can be computed from year and JD)

JD

Day of the year ([1, 365]; numeric)

maxTemp

Maximum temperature (degrees Celsius; numeric).

minTemp

Minimum temperature (degrees Celsius; numeric).

SolarR

Solar radiation (MJ / m^2; numeric).

PotEv

Potential evaporation (kg / m^2 ~ mm; numeric)

precip

Precipitation (kg / m^2 ~ mm; numeric)

Source

http://www.noaa.gov/

Examples

data(WD1979)
summary(WD1979)

#>       year       month           day                JD         maxTemp      
#>  Min.   :1979   Mode:logical   Mode:logical   Min.   :  1   Min.   :-10.60  
#>  1st Qu.:1979   NA's:365       NA's:365       1st Qu.: 92   1st Qu.: 11.50  
#>  Median :1979                                 Median :183   Median : 21.66  
#>  Mean   :1979                                 Mean   :183   Mean   : 18.91  
#>  3rd Qu.:1979                                 3rd Qu.:274   3rd Qu.: 27.78  
#>  Max.   :1979                                 Max.   :365   Max.   : 33.19  
#>     minTemp             SolarR           PotEv             precip      
#>  Min.   :-22.3562   Min.   : 1.650   Min.   :-0.0790   Min.   : 0.000  
#>  1st Qu.:  0.2715   1st Qu.: 6.744   1st Qu.: 0.2840   1st Qu.: 0.000  
#>  Median :  9.5810   Median : 9.694   Median : 0.5760   Median : 0.000  
#>  Mean   :  8.0733   Mean   : 9.751   Mean   : 0.5498   Mean   : 2.601  
#>  3rd Qu.: 17.0856   3rd Qu.:12.882   3rd Qu.: 0.8030   3rd Qu.: 2.600  
#>  Max.   : 24.6433   Max.   :16.753   Max.   : 1.0800   Max.   :42.900