Champaign, IL weather data

Multiple years of Champaign, IL weather data specified at hourly intervals in the CST time zone. The columns correspond to typical inputs required for BioCro simulations. Note: some values are missing near the start of 1995 since the data is not available from SURFRAD.

This weather data is included in the BioCro package so users can reproduce calculations in the BioCro II paper (citation forthcoming) and for exploratory purposes; it is likely that most BioCro studies will require different data sets, and no attempt is made here to be exhaustive.

## YYYY should be replaced with a 4-digit year from the available set of years,
## e.g. 2002
weatherYYYY

Format

26 data frames corresponding to 26 years of weather data (1995 - 2020), where each data frame has 8760 or 8784 rows (representing hourly time points) and 9 columns:

year: the year

doy: the day of year

hour: the hour

time_zone_offset: the time zone offset relative to UTC (hr)

precip: preciptation rate (mm / hr)

rh: the ambient relative humidity (dimensionless)

solar: the incoming photosynthetically active photon flux density (PPFD) measured on a ground area basis including direct and diffuse sunlight light just outside the crop canopy (micromol / m^2 / s)

temp: the ambient air temperature (degrees Celsius)

windspeed: the wind speed in the ambient air just outside the canopy (m / s)

Source

Weather data were obtained from the public SURFRAD and WARM databases and processed according to the method described in the BioCro II paper (submitted). See that paper for a full description of the data processing.

In brief, the columns in the data frames were determined from SURFRAD and WARM variables as follows:

precip: from the precip variable in the WARM data set

rh: from the rh variable in the SURFRAD data set

solar: from the par variable in the SURFRAD data set; when these values are not available, the direct_n, diffuse, and zen variables are used to make an estimate

temp: from the temp variable in the SURFRAD data set

windspeed: from the windspd variable in the SURFRAD data set

The WARM data set includes daily values. Hourly values for precipitation are derived from daily totals by assuming a constant rate of precipitation throughout the day.

The SURFRAD data set includes values at 1 or 3 minute intervals. Hourly values are determined by averaging over hourly intervals, where the value at hour h is the average over the hour-long interval centered at h. Some values are missing; any missing entries are filled via an interpolation procedure based on the assumption that values at the same hour of sequential days should be similar.

The time_zone_offset is set to a constant value of -6 since this data is specified in the CST time zone (i.e., UTC-6). Since the value of this quantity does not change, it could in principle be considered a parameter rather than a driver; however, it is included with the weather data for convenience.