Crop specifications

Sets of initial values, parameters, direct modules, differential modules, and differential equation solvers that can be used (along with the included weather data) to run growth simulations for several crops in Champaign, IL. For some of these crops, the parameter and module specifications are associated with published papers:

glycine_max (unpublished)

manihot_esculenta (unpublished)

miscanthus_x_giganteus

sorghum (unpublished)

soybean

willow

zea_mays (unpublished)

The soybean specifications are intended to be used along with the special soybean weather data (cmi_soybean_weather_data). The other crops should be used with the other weather data (cmi_weather_data).

Important note: Some specifications, such as the values of photosynthetic parameters, would remain the same in any location; others, such as the latitude or longitude, would need to change when simulating crop growth in different locations. Care must be taken to understand each input quantity before attempting to run simulations in other places.

Usage

## CROP should be replaced by the name of a valid crop, e.g. `soybean`.
## SPECIFICATION should be replaced by a specification type, e.g. `parameters`.
CROP_SPECIFICATION

Format

Each crop has associated initial values, parameters, direct modules, differential modules, and differential equation solvers. These objects are prefixed with the name of the crop; for example, the soybean parameters are called soybean_parameters. In general, the specifications have the following format:

CROP_initial_values: a list of named elements specifying the initial values of differential quantities for crop CROP

CROP_parameters: a list of named elements specifying the values of the parameters for crop CROP

CROP_direct_modules: a list of direct modules required for crop CROP; some elements are named to facilitate swapping out different modules that can fulfill the same role

CROP_differential_modules: a list of differential modules required for crop CROP; some elements are named to facilitate swapping out different modules that can fulfill the same role

CROP_ode_solver: a list of named elements that specify the type of solver to use for crop CROP, as well as any additional parameters that determine its behavior

The soybean model has additional lists of initial values, parameters, direct modules, and differential modules that can be used for its circadian clock component. These lists are prefixed by soybean_clock.

Source

For the published crop models, see the following papers:

Miguez, F. E., Zhu, X., Humphries, S., Bollero, G. A. & Long, S. P. A semimechanistic model predicting the growth and production of the bioenergy crop Miscanthus x giganteus: description, parameterization and validation. GCB Bioenergy 1, 282-296 (2009).

Miguez, F. E., Maughan, M., Bollero, G. A. & Long, S. P. Modeling spatial and dynamic variation in growth, yield, and yield stability of the bioenergy crops Miscanthus x giganteus and Panicum virgatum across the conterminous United States. GCB Bioenergy 4, 509-520 (2012).

Wang, D. et al. A physiological and biophysical model of coppice willow (Salix spp.) production yields for the contiguous USA in current and future climate scenarios. Plant, Cell & Environment 38, 1850-1865 (2015).

Matthews, M. et al. Soybean-BioCro: A semi-mechanistic model of soybean growth. in silico Plants (2021).