BioCro Module Functions

module_info returns essential information about a BioCro module.

evaluate_module runs a BioCro module using a list of input quantity values.

quantity_list_from_names initializes a list of named numeric elements from a set of names

module_info(module_name, verbose = TRUE)

  evaluate_module(module_name, input_quantities)

  quantity_list_from_names(quantity_names)

Arguments

module_name
A string representing the name of a BioCro module; only one name is allowed at the moment, but multiple modules will be supported in the future

verbose
A boolean indicating whether or not to print information to the R console

input_quantities
A list of named numeric elements representing the input quantities required by the module; any extraneous quantities will be ignored by the module

quantity_names
A vector or list of strings

Details

By providing avenues for retrieving information about a module and evaluating a module's equations, the module_info and evaluate_module functions form the main interface to individual BioCro modules from within R. The quantity_list_from_names function is a convenience function for preparing suitable quantity lists to pass to evaluate_module.

Value

module_info

An invisible list of several named elements containing essential information about the module:

module_name: The module's name

inputs: A character vector of the module's inputs

outputs: A character vector of the module's outputs

type: The module's type represented as a string (either 'differential' or 'direct')

euler_requirement: Indicates whether the module requires a fixed-step Euler ODE solver when used in a BioCro simulation

creation_error_message: Describes any errors that occurred while creating an instance of the module

evaluate_module

A list of named numeric elements representing the values of the module's outputs as calculated from the input_quantities according to the module's equations

quantity_list_from_names

A list of named numeric elements, where the names are set by quantity_names and each value is initialized to 1.

See also

Examples

# Example 1: printing information about the 'c3_assimilation' module to the R
# console
module_info('c3_assimilation')
#> 
#> 
#> Module name:
#>   c3_assimilation
#> 
#> Module input quantities:
#>   Qp
#>   Tleaf
#>   rh
#>   vmax1
#>   jmax
#>   tpu_rate_max
#>   Rd
#>   b0
#>   b1
#>   Gs_min
#>   Catm
#>   atmospheric_pressure
#>   O2
#>   theta
#>   StomataWS
#>   water_stress_approach
#>   electrons_per_carboxylation
#>   electrons_per_oxygenation
#> 
#> Module output quantities:
#>   Assim
#>   Gs
#>   Ci
#>   GrossAssim
#> 
#> Module type (differential or direct):
#>   direct
#> 
#> Requires a fixed step size Euler ode_solver:
#>   no
#> 

# Example 2: getting the inputs to the 'thermal_time_linear' module, generating
# a default input list, and using it to run the module
info <- module_info('thermal_time_linear', verbose = FALSE)
inputs <- quantity_list_from_names(info$inputs) # All inputs will be set to 1
outputs <- evaluate_module('thermal_time_linear', inputs)