Complete Miscanthus assimilation field data

Assimilation and stomatal conductance in Miscanthus as measured in Beale, Bint, and Long 1996. (Missing data are also included.) The first column is the date, the second the hour. Columns 3 and 4 are assimilation and stomatal conductance respectively.

Format

Data frame of dimensions 35 by 6.

Source

C. V. Beale, D. A. Bint, S. P. Long. 1996. Leaf photosynthesis in the C4-grass Miscanthus x giganteus, growing in the cool temperate climate of southern England. J. Exp. Bot. 47 (2): 267--273.

Details

The third column is the observed net assimilation rate (micromoles m^-2 s^-1).

The fifth column is the observed quantum flux (micromoles m^-2 s^-1).

The sixth column is the temperature (degrees Celsius).