Convenience Functions for Partial Application

Convenience functions for using partial application with BioCro

partial_run_biocro(
    initial_values = list(),
    parameters = list(),
    drivers,
    direct_module_names = list(),
    differential_module_names = list(),
    ode_solver = default_ode_solver,
    arg_names,
    verbose = FALSE
)

Details

Partial application is the technique of fixing some of the input arguments to a function, producing a new function with fewer inputs. In the context of BioCro, partial application can often be useful while varying some parameters, initial values, or drivers while performing optimization or sensitivity analysis.

Value

A function that calls run_biocro with all of the inputs (except those specified in 'arg_names') set as their default values. The only input argument of the new function is a numerical vector specifying the values of the quantities in 'arg_names'. In this vector, the values must be supplied in the same order as in 'arg_names', and arguments corresponding to drivers must be vectors having the same length as the other drivers.

See also

Examples

# Example: varying the thermal time values at which senescence starts for
# different organs in a simulation; here we set them to the following values
# instead of the defaults:
#  - seneLeaf: 2000 degrees C * day
#  - seneStem: 2100 degrees C * day
#  - seneRoot: 2200 degrees C * day
#  - seneRhizome: 2300 degrees C * day
senescence_simulation <- partial_run_biocro(
    sorghum_initial_values,
    sorghum_parameters,
    get_growing_season_climate(weather2005),
    sorghum_direct_modules,
    sorghum_differential_modules,
    sorghum_ode_solver,
    c('seneLeaf', 'seneStem', 'seneRoot', 'seneRhizome')
)
result <- senescence_simulation(c(2000, 2100, 2200, 2300))

# Example: a crude method for simulating the effects of climate change; here we
# increase the atmospheric CO2 concentration to 500 ppm and the temperature by
# 2 degrees C relative to 2002 temperatures. The two commands below are
# equivalent, so `hot_result_1` is identical to `hot_result_2`.
temperature_simulation <- partial_run_biocro(
  soybean_initial_values,
  soybean_parameters,
  soybean_weather2002,
  soybean_direct_modules,
  soybean_differential_modules,
  soybean_ode_solver,
  c("Catm", "temp")
)
hot_result_1 <- temperature_simulation(c(500, soybean_weather2002$temp + 2.0))
hot_result_2 <- temperature_simulation(list(Catm = 500, temp = soybean_weather2002$temp + 2.0))