Simultaneous equations

These functions are only intended for use by Ed. This documentation is only here to avoid warnings from R CMD check and is not intended to actually describe what these functions do or how they should be used.

solve_simultaneous_equations(
    direct_modules,
    known_quantities,
    unknown_quantities,
    lower_bounds,
    upper_bounds,
    abs_error_tols,
    rel_error_tols,
    max_it = 50,
    solver_type = "newton_raphson_backtrack_boost",
    silent = FALSE
  )

  validate_simultaneous_equations(
    direct_modules,
    known_quantities,
    unknown_quantities,
    silent = FALSE
  )

  test_simultaneous_equations(
    direct_modules,
    known_quantities,
    unknown_quantities
  )

  get_all_se_solvers()

Arguments

abs_error_tols

Please ignore this.

direct_modules

Please ignore this.

known_quantities

Please ignore this.

lower_bounds

Please ignore this.

max_it

Please ignore this.

rel_error_tols

Please ignore this.

silent

Please ignore this.

solver_type

Please ignore this.

unknown_quantities

Please ignore this.

upper_bounds

Please ignore this.

Value

solve_simultaneous_equations

Please ignore this.

validate_simultaneous_equations

Please ignore this.

test_simultaneous_equations

Please ignore this.

get_all_se_solvers

Please ignore this.