Calculate Derivatives for Differential Quantities

Solving a BioCro model using one of R's available differential equation solvers

Usage

system_derivatives(
  parameters = list(),
  drivers,
  direct_module_names = list(),
  differential_module_names = list()
)

Arguments

The input arguments to system_derivatives are the same as the corresponding inputs to run_biocro.

Value

The return value of system_derivatives is a function with three inputs (t, differential_quantities, and parms) that returns derivatives for each of the differential quantities in the dynamical system determined by the original inputs (parameters, drivers,

direct_module_names, and differential_module_names).

This function signature and the requirements for its inputs are set by the

LSODES function from the deSolve package. The t input should be a single time value and the differential_quantities input should be a vector with the names of the differential quantities defined by the modules. parms is required by LSODES, but we don't use it for anything.

This function can be passed to LSODES as an alternative integration method, rather than using one of BioCro's built-in solvers.

See also

Examples

# Example 1: calculating a single derivative using a soybean model

soybean_system <- system_derivatives(
  soybean_parameters,
  soybean_weather2002,
  soybean_direct_modules,
  soybean_differential_modules
)

derivs <- soybean_system(0, unlist(soybean_initial_values), NULL)

# Example 2: a simple oscillator with only one module (requires deSolve)

if (FALSE) {
if (require(deSolve)) {

  times = seq(0, 5, length=100)

  oscillator_system_derivatives <- system_derivatives(
    list(
      timestep = 1,
      mass = 1,
      spring_constant = 1
    ),
    data.frame(time=times),
    character(),
    'harmonic_oscillator'
  )

  result <- as.data.frame(lsodes(
    c(position=0, velocity=1),
    times,
    oscillator_system_derivatives
  ))

  lattice::xyplot(
    position + velocity ~ time,
    type='l',
    auto=TRUE,
    data=result
  )
}
}

# Example 3: solving 500 hours of a soybean simulation. This requires the
# deSolve package and will run very slow compared to a regular call to
# `run_biocro`.
if (FALSE) {
if (require(deSolve)) {
  soybean_system <- system_derivatives(
    soybean_parameters,
    soybean_weather2002,
    soybean_direct_modules,
    soybean_differential_modules
  )

  times = seq(from=0, to=500, by=1)

  result <- as.data.frame(lsodes(unlist(soybean_initial_values), times, soybean_system))

  lattice::xyplot(Leaf + Stem ~ time, type='l', auto=TRUE, data=result)
}
}