Downscales daily weather to hourly

Manipulates weather data in the format obtained from WARM (see link below) and returns the format and units needed for most functions in this package. This function should be used for one year at a time. It returns hourly (or sub-daily) weather information.

weach(X, ...)

Arguments

X
a matrix (or data frame) containing weather information. The input format is strict but it is meant to be used with the data usually obtained from weather stations in Illinois. The data frame should have 11 columns (see details).

latitude
latitude at the specific location

ts
timestep for the simulation of sub-daily data from daily. For example a value of 3 would return data every 3 hours. Only divisors of 24 work (i.e. 1, 2, 3, 4, etc.).

temp.units
Option to specify the units in which the temperature is entered. Default is Farenheit.

rh.units
Option to specify the units in which the relative humidity is entered. Default is percent.

ws.units
Option to specify the units in which the wind speed is entered. Default is miles per hour.

pp.units
Option to specify the units in which the precipitation is entered. Default is inches.

list()
additional arguments to be passed to lightME

Value

a matrix returning hourly (or sub-daily) weather data. Dimensions 8760 (if hourly) by 8.

year Year.

doy Day of the year.

hour Hour of the day ([0, 23], depending on the timestep).

solar Direct solar radiation (micromoles m^-2 s^-1).

temp Air temperature (degrees Celsius).

rh Relative humidity ([0, 1]).

windspeed Average wind speed (m s^-1).

precip Precipitation (mm)

Details

This function was originally used to transform daily data to hourly data. Some flexibility has been added so that other units can be used. The input data used originally looked as follows.

col 1 year

col 2 day of the year ([1, 365]). Does not consider leap years.

col 3 total daily solar radiation (MJ / m^2).

col 4 maximum temperature (degrees Fahrenheit).

col 5 minimum temperature (degrees Fahrenheit).

col 6 average temperature (degrees Fahrenheit).

col 7 maximum relative humidity (%).

col 8 minimum relative humidity (%).

col 9 average relative humidity (%).

col 10 average wind speed (mi / hr).

col 11 precipitation (inches).

All the units above are the defaults but they can be changed as part of the arguments.

Author

Fernando E. Miguez.

Examples