Run module test cases for an entire BioCro module library

Modules can be tested using test cases, which are sets of known outputs that correspond to particular inputs. The test_module_library function provides a way to run all test cases for all modules in a BioCro module library.

Usage

test_module_library(library_name, directory, modules_to_skip = c())

Arguments

library_name

The name of a BioCro module library.

directory

The directory where module test case files are stored, e.g. file.path('tests', 'module_test_cases')

modules_to_skip

A vector of local module name strings indicating any modules from the library that should not be tested. This feature should be used sparingly, since there are very few legitimate reasons to skip a module test.

Details

For each module in the specified library, test_module_library loads stored test cases from the specified directory and runs each test case, storing information about any test failures or other issues that may occur. If any problems are detected, test_module_library throws an error with a message describing the issues.

For an example of how this function can be used along with the testthat package, see tests/testthat/test.Modules.R.

Value

None

See also

Examples

# Here we will initialize a module test case file in a temporary directory, and
# then use `test_module_library` to test it. We will need to skip most of the
# modules in the library, since we only have a test case for one of them.

td <- tempdir()

initialize_csv(
  'BioCro:thermal_time_linear',
  td,
  nonstandard_inputs = list(temp = -1),
  overwrite = TRUE
)
#> [1] "Case file `/tmp/RtmpnV0EZU/BioCro_thermal_time_linear.csv` was initialized; any pre-existing file was overwritten"

# Get a list of local module names, excluding the module that has a test case
all_modules <- get_all_modules('BioCro')
skip <- all_modules[all_modules != 'BioCro:thermal_time_linear']
skip <- gsub('BioCro:', '', skip)

test_module_library('BioCro', td, skip)

# If we attempt to test the entire library, we will get errors since only one
# module actually has an associated case file
tryCatch(
  {
    test_module_library('BioCro', td)
  },
  error = function(e) {print(e)}
)
#> <simpleError in test_module_library("BioCro", td): Problems occurred while testing modules from the `BioCro` library:
#>   Module `BioCro:aba_decay`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_aba_decay.csv` does not exist
#>   Module `BioCro:ball_berry`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_ball_berry.csv` does not exist
#>   Module `BioCro:biomass_leaf_n_limitation`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_biomass_leaf_n_limitation.csv` does not exist
#>   Module `BioCro:buck_swvp`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_buck_swvp.csv` does not exist
#>   Module `BioCro:bucket_soil_drainage`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_bucket_soil_drainage.csv` does not exist
#>   Module `BioCro:c3_assimilation`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_c3_assimilation.csv` does not exist
#>   Module `BioCro:c3_canopy`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_c3_canopy.csv` does not exist
#>   Module `BioCro:c3_leaf_photosynthesis`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_c3_leaf_photosynthesis.csv` does not exist
#>   Module `BioCro:c4_assimilation`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_c4_assimilation.csv` does not exist
#>   Module `BioCro:c4_canopy`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_c4_canopy.csv` does not exist
#>   Module `BioCro:c4_leaf_photosynthesis`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_c4_leaf_photosynthesis.csv` does not exist
#>   Module `BioCro:development_index`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_development_index.csv` does not exist
#>   Module `BioCro:example_model_mass_gain`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_example_model_mass_gain.csv` does not exist
#>   Module `BioCro:example_model_partitioning`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_example_model_partitioning.csv` does not exist
#>   Module `BioCro:fake_solar`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_fake_solar.csv` does not exist
#>   Module `BioCro:FvCB`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_FvCB.csv` does not exist
#>   Module `BioCro:golden_ratio_hyperbola`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_golden_ratio_hyperbola.csv` does not exist
#>   Module `BioCro:grimm_soybean_flowering`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_grimm_soybean_flowering.csv` does not exist
#>   Module `BioCro:grimm_soybean_flowering_calculator`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_grimm_soybean_flowering_calculator.csv` does not exist
#>   Module `BioCro:harmonic_energy`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_harmonic_energy.csv` does not exist
#>   Module `BioCro:harmonic_oscillator`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_harmonic_oscillator.csv` does not exist
#>   Module `BioCro:hyperbola_2d`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_hyperbola_2d.csv` does not exist
#>   Module `BioCro:incident_shortwave_from_ground_par`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_incident_shortwave_from_ground_par.csv` does not exist
#>   Module `BioCro:leaf_gbw_nikolov`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_leaf_gbw_nikolov.csv` does not exist
#>   Module `BioCro:leaf_gbw_thornley`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_leaf_gbw_thornley.csv` does not exist
#>   Module `BioCro:leaf_shape_factor`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_leaf_shape_factor.csv` does not exist
#>   Module `BioCro:leaf_water_stress_exponential`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_leaf_water_stress_exponential.csv` does not exist
#>   Module `BioCro:light_from_solar`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_light_from_solar.csv` does not exist
#>   Module `BioCro:linear_vmax_from_leaf_n`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_linear_vmax_from_leaf_n.csv` does not exist
#>   Module `BioCro:litter_cover`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_litter_cover.csv` does not exist
#>   Module `BioCro:magic_clock`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_magic_clock.csv` does not exist
#>   Module `BioCro:Module_1`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_Module_1.csv` does not exist
#>   Module `BioCro:Module_2`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_Module_2.csv` does not exist
#>   Module `BioCro:Module_3`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_Module_3.csv` does not exist
#>   Module `BioCro:night_and_day_trackers`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_night_and_day_trackers.csv` does not exist
#>   Module `BioCro:no_leaf_resp_neg_assim_partitioning_growth_calculator`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_no_leaf_resp_neg_assim_partitioning_growth_calculator.csv` does not exist
#>   Module `BioCro:no_leaf_resp_partitioning_growth_calculator`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_no_leaf_resp_partitioning_growth_calculator.csv` does not exist
#>   Module `BioCro:nr_ex`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_nr_ex.csv` does not exist
#>   Module `BioCro:one_layer_soil_profile`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_one_layer_soil_profile.csv` does not exist
#>   Module `BioCro:one_layer_soil_profile_derivatives`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_one_layer_soil_profile_derivatives.csv` does not exist
#>   Module `BioCro:oscillator_clock_calculator`: could not load test cases: Error in cases_from_csv(module, directory): Module test case file `/tmp/RtmpnV0EZU/BioCro_oscillator_clock_calculator.csv` does not exist
#>   Module `BioCro:parameter_calculator`: could not load test cases: Error in cases_from_csv(modu>